Add Dockerfile
Dockerfile is meant to be a complete boxed env for compiling rust-wiiu code. Once this is more thoroughly tested, it might become the baseline for something to be considered "compilable" across rust-wiiu (i.e. if something compiles on one machine but not on the container then it is considered broken; something like this)